Cleanup unnecessary code and move smaller files together (performance)
